Minimal fix for Drone in 1.13
The deprecated Block.setTypeIdAndData method that the Drone was using
was finally removed from the Bukkit API as part of the 1.13 update.
There is no longer a method on the Block interface that can work with
the old numeric block IDs.

Fortunately, CraftBukkit/Spigot now has a `CraftEvil` class offering
a static `setTypeIdAndData` method that maps the old IDs to the new
API. It is definitely not future-proof but allows us to run on 1.13
until we come up with a way to use the more recent API without breaking
a lot of existing ScriptCraft code.

This patch simply detects when the `Block.setTypeIdAndData` is missing
and tries to use the `CraftEvil` class instead.